cleanup for playback functions
